+++ main.go
@@ -6,23 +6,17 @@ import (
 	"flag"
 	"fmt"
 	"log"
-	"os/exec"
-	"time"
+	"os"
+	"unsafe"
 )
 
-// http://technosophos.com/2014/06/11/compile-time-string-in-go.html
-// go build -ldflags "-x main.version $(git describe --tags)"
 var version string
 var name string = "refresh"
 
 var showVersion bool
-var panelTitle string
-var sendDelay time.Duration
 
 func init() {
 	flag.BoolVar(&showVersion, "version", false, "print version and exit")
-	flag.StringVar(&panelTitle, "title", "Environment Variables", "localized version of title")
-	flag.DurationVar(&sendDelay, "delay", 100*time.Millisecond, "Delays in milliseconds before sending VK_RETURN")
 }
 
 func main() {
@@ -32,13 +26,15 @@ func main() {
 	if showVersion {
 		fmt.Printf("%s version %s\n", name, version)
 	} else {
-		cmd := exec.Command("cmd", "/c", "start rundll32 sysdm.cpl,EditEnvironmentVariables")
-		err := cmd.Run()
-		if err != nil {
-			log.Fatal("Failed executing command:", err)
+		ret := SendMessageTimeout(HWND_BROADCAST, WM_SETTINGCHANGE, 0,
+			uintptr(unsafe.Pointer(StringToUTF16Ptr("Environment"))), SMTO_ABORTIFHUNG, 5000)
+
+		// If the function succeeds, the return value is nonzero.
+		if ret == 0 {
+			fmt.Println("Refresh: Error")
+			os.Exit(1)
+		} else {
+			fmt.Println("Refresh: Success")
 		}
-		time.Sleep(sendDelay)
-		findWindow(panelTitle)
-		sendKey(VK_RETURN)
 	}
 }